+++ packages/af-utils-react-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @af-utils/react-table
 
+## 0.0.26
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.0.21
+
 ## 0.0.25
 
 ### Patch Changes